Cleanup and added layers
Moved from working in a single script to actually building the package.
Currently the non-linear Flux layer kind of works, but needs further testing.
Combining this layer with Dense layers still has issues with SVector multiplication.
Neither can be used with params(), but this needs to be fixed.
FinnisSinclair.jl model remains unchanged, but will eventually be deprecated for these layers.