added tests and cleanup
Implemented an energy and forces functions natively in ACEgnns. Implemented a test for derivative of energy and it works.
Test for forces still needs work.
Derivative of forces still, not implemented. Requires a smart rrule that doesn't go through all that mutation.
For testing this last part we will need to wait for ACE to finish properly testing grad_params_configs and then pushing the new changes.